eslint fixes, move a few things
