+++ src/routes/GamePlay/utils.ts
@@ -1,0 +1,13 @@
+import { GBSetupSteps } from "../../models/gbdb";
+
+
+export function stepToNav(step: GBSetupSteps) {
+  switch (step) {
+    case "Guilds":
+      return "/game";
+    case "Draft":
+      return "/game/draft";
+    case "Game":
+      return "/game/draft/play";
+  }
+}
